pyproject.toml
[project]
name = "gardening-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "requests",
    "anthropic",
    "scikit-learn",
    "numpy",
    "pandas",
    "beautifulsoup4",
    "pdfplumber",
    "chromadb",
    # FastAPI backend
    "fastapi",
    "uvicorn[standard]",
    "alembic",
    "python-multipart",
    "sqlalchemy",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
eda = ["matplotlib", "seaborn", "jupyter"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = false
